requirements.txt
streamlit>=1.32.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
plotly>=5.18.0
anthropic>=0.40.0
praw>=7.7.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
